Isolating Physics engine and fixing hover on link effect
