+++ src/main.ts
@@ -3,26 +3,18 @@ import * as THREE from "three";
 import SetScene from "./SetScene";
 import SetListeners from "./SetListeners";
 
-// Utility function to map a value from one range to another
-const map = (
-  value: number,
-  min1: number,
-  max1: number,
-  min2: number,
-  max2: number
-): number => {
-  return min2 + ((value - min1) * (max2 - min2)) / (max1 - min1);
-};
+import {
+  sendPhysicsUpdate,
+  physicsWorker,
+  setCreate,
+  numberOfSpheres,
+  geometryData,
+} from "./PhysicsUtils";
 
-const dt = 1 / 60;
-export const numberOfSpheres = Math.round(
-  map(window.innerWidth, 300, 2000, 5, 30)
-);
+export let spheres = []; // Array to hold the sphere objects
 
 let timeSinceLast = 0;
 let maxTime = 5; // Seconds in between sphere creation
-
-let spheres = []; // Array to hold the sphere objects
 
 // Check if device is touchscreen
 export const isOnTouchScreen = "ontouchstart" in window;
@@ -31,18 +23,6 @@ export const isOnTouchScreen = "ontouchstart" in window;
 export const container = document.querySelector(
   ".canvas-container"
 ) as HTMLElement;
-
-// Data containers for physics and geometry information
-let physicsData = {
-  positions: new Float32Array(numberOfSpheres * 3),
-  quaternions: new Float32Array(numberOfSpheres * 4),
-  scales: new Float32Array(numberOfSpheres * 4),
-};
-export const geometryData = {
-  positions: new Float32Array(numberOfSpheres * 3),
-  quaternions: new Float32Array(numberOfSpheres * 4),
-  scales: new Float32Array(numberOfSpheres * 4),
-};
 
 // Set up the Scene using the setScene function
 export var {
@@ -56,7 +36,6 @@ export var {
   mouse,
   mouseTarget,
   mouseScaleTarget,
-  mosueOverLink,
   move,
   tmpM,
   offset,
@@ -69,64 +48,12 @@ export var {
   geometry,
 } = SetScene();
 
-// Start physics engine/worker
-const physicsWorker = new Worker("/src/worker-physics.js");
-
-let sendTime;
-let create = false;
-let needsupdate = true;
-
-// Event handler for receiving messages from the physics worker
-physicsWorker.onmessage = function (e) {
-  // Update physics data with the received positions, quaternions, and scales
-  physicsData.positions = e.data.positions;
-  physicsData.quaternions = e.data.quaternions;
-  physicsData.scales = e.data.scales;
-
-  // Update geometry data with the physics data
-  geometryData.positions.set(physicsData.positions);
-  geometryData.quaternions.set(physicsData.quaternions);
-  geometryData.scales.set(physicsData.scales);
-
-  needsupdate = true;
-  // Schedule the next update based on the time difference
-  setTimeout(
-    UpdatePhysicsWorker,
-    Math.max(dt * 1000 - (Date.now() - sendTime), 0)
-  );
-};
-
-// Function to send update to the physics worker
-const UpdatePhysicsWorker = () => {
-  if (!needsupdate) {
-    return;
-  }
-  needsupdate = false;
-
-  // Record the current time and send the updated data to the physics worker
-  sendTime = Date.now();
-  physicsWorker.postMessage(
-    {
-      create: create,
-      numberOfSpheres: spheres.length,
-      dt: dt,
-      positions: physicsData.positions,
-      quaternions: physicsData.quaternions,
-      scales: physicsData.scales,
-      mouse: move,
-    },
-    [
-      physicsData.positions.buffer,
-      physicsData.quaternions.buffer,
-      physicsData.scales.buffer,
-    ]
-  );
-  create = false;
-};
-
-
 // Animation loop
 export var scrollPercent = 0;
+export const mouseState = {
+  mouseOverLink: false,
+};
+
 const animate = () => {
   time += 0.01;
   timeSinceLast++;
@@ -142,7 +69,7 @@ const animate = () => {
 
   // Create new spheres at regular intervals
   if (spheres.length < numberOfSpheres && timeSinceLast > maxTime) {
-    create = true;
+    setCreate(true);
     spheres.push(1);
     timeSinceLast = 0;
   }
@@ -189,9 +116,14 @@ const animate = () => {
 
   // Scale and rotate the mouseBall based on mouse interactions
   mouseScaleTarget.setScalar(
-    mosueOverLink ? 0.25 + mouse.distanceToSquared(mouseTarget) * 0.5 : 0.1
+    mouseState.mouseOverLink
+      ? 0.25 + mouse.distanceToSquared(mouseTarget) * 0.5
+      : 0.1
   );
   mouseBall.rotateZ(-0.2 * mouseBall.scale.x);
+
+  mouseBall.rotateZ(-0.2 * mouseBall.scale.x);
+
   mouseBall.scale.lerp(mouseScaleTarget, 0.06);
 
   // Render the scene
@@ -199,7 +131,7 @@ const animate = () => {
 };
 
 // Start the update process for the physics worker
-UpdatePhysicsWorker();
+sendPhysicsUpdate(physicsWorker);
 
 // Start the animation loop
 animate();
